refactor: restaurantEnrollModal 내부의 로직 분리
